Progress entity, repo, service and tests added, need to be refactore because of new table Hint

+++ src/test/java/dk/cngroup/lentils/repository/TeamRepositoryTest.java
@@ -56,7 +56,7 @@ public class TeamRepositoryTest {
 
         List<Team> teamsFound = repository.findAll();
 
-        assertEquals((int)generator.NUMBER_OF_TEAMS, teamsFound.size());
+        assertEquals(generator.NUMBER_OF_TEAMS, teamsFound.size());
     }
 
     @Test
